server/requirements.txt
certifi==2018.8.24
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.6
gcm==0.3
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
nltk==3.3
pyenchant==2.0.0
requests==2.19.1
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
